animation: Add zoom animation
This is a simple affine transformation that goes from
one box to another.

The animation runs "from" some source point "to" some target point
which is the surface' "natural" position within the scene (eg, if
there were a scene graph, the "to" position forms its co-ordinates
and its geometry).

For instance, if the window was minimizing, then the animation would
have to be run in reverse, "from" the minimized point "to" the unminimized
point, but stepping backwards.