animation-glib: Add constructor helpers
These helpers are used to quickly append a new "interface"
construction prop to the existing construction properties
and make it easy to lookup the props inline.

Basically, we need these helpers because we are wrapping some
C++ interface implementation in a GObject. GObject doesn't require
all the properties to be set at construction time, whereas
constructing the C++ interface implementation might. So we need to
collect all the properties in the constructor and then construct
the underlying C++ interface there, so that it is immediately
available to the base class as a property on the object.

Doing this by looking over all the construct properties and assigning
values would be cumbersome, especially doing it for every effect
implementation. Instead, we put all the construct properties into
a hash table (hashed by the property name) and then fetch them in
O(1) using the InterfaceConstructor template.

This template constructs the "Interface" type (the type passed
should be the implementation of the interface you want to construct)
and variadically takes in all the construction parameters. Use
ForwardValueFromHT to extract a typed value from the given property name
from its corresponding GValue in the construct properties hash table. For
instance, supposing that MyType's constructor is defined as follows:

MyType::MyType (float                   bar,
                animation::Point const &baz)
{
  ...
}

You would use the following in the constructor override for the GObject
wrapper:

const char *interesting_props = {
  "bar",
  "baz",
  NULL
};
g_autoptr(GHashTable) props_ht =
  static_hash_table_of_values_for_specs (interesting_props,
                                         construct_params,
                                         n_construct_params);
auto *interface = InterfaceConstructor <MyType>::construct (
  ForwardValueFromHT (props_ht, g_value_get_float, "bar"),
  ForwardValueFromHT (props_ht, animation_point_from_gvalue, "baz")
);
g_auto(GValue) interface_value = G_VALUE_INIT;
unsigned int new_n_construct_params = 0;
GObjectConstructParam *new_construct_params =
  append_interface_prop_to_construct_params (construct_params,
                                             n_construct_params,
                                             object_class,
                                             &interface_value,
                                             interface,
                                             &new_n_construct_params);

object_class->constructor (type,
                           new_construct_params,
                           new_n_construct_params);

Another thing you may wish to do is ensure that certain properties
always take a certain value that is not their default (for instance,
you might want to ensure tha a certain object-valued property
is some non-NULL sane default in the event that it is not set). To
do that, you can use replace_named_pointer_props_in_construct_params_if_null,
which will only construct the desired placeholder object property values
in the event that the property provided by the user or by default
due to the G_PARAM_SPEC_CONSTRUCT flag was NULL. For example:

  /* We need to also set defaults for certain properties in order
   * to ensure that they are bindable later on, in case they are
   * not set by the caller. */
  replace_named_pointer_props_in_construct_params_if_null (
    construct_params,
    n_construct_params,
    {
      ReplacePropSpec ("stepper", g_value_get_object, g_value_set_object, []() -> gpointer {
        return animation_linear_stepper_new (1);
      }),
      ReplacePropSpec ("target", g_value_get_object, g_value_set_object, []() -> gpointer {
        return animation_box_query_new ();
      })
    }
  );

Note that with most of the animations in this library, you can get away with
the latter only - which is to say that most of the animations can be default-constructed,
however their default properties may not be easily bindable because they refer to some
object that is not being wrapped by a GObject that can be exposed to the introspection
system. Therefore, the animation author's job is to ensure that the properties
set on the underlying C++ type support bindability.