animation: Add box_calculation inline functions
We'll use these later for the affine transformation animations
to calculate box offsets.

Usually we want to compute where the center of a box is in both
absolute co-ordinates and relative co-ordinates. ComputeBoxCenter
will do it in absolute co-ordinates and ComputeBoxCenterOffset
in co-ordinates relative to the top left corner of the box.