animation: Add animation::BoxQuery interface
When animating from one position to another, it is possible that
some things could change underneath us. For instance, the target position
of the surface may be changed by the user while the animation is in progress.

In this case, the animation needs to adapt to this change by allowing
the caller some mechanism to transparently change the target box
tha the surface is animating to. This can be done by overriding
the animation::BoxQuery class and calling the Update() method as
appropriate, which will update the saved geometry of the target position.

For example, in a windowing system, one might set up a listener
for surface geometry events in their animation::BoxQuery subclass,
which then calls Update() when the geometry has been changed. This
obviates the need for the caller to explicitly update all animations.