animation-clutter: Add animation-clutter library and common helpers
This library will be used to bind libanimation to Clutter, so that
it can be used in GNOME-Shell. animation-clutter-common is a
set of helper functions that are used to translate boxes into the
correct space (eg, expanding paint boxes to cover the FBO size
for a deform effect, determining the paint box extents of an
actor, etc).