animation-clutter: Add AnimationClutterActorBoxQuery
This class extends AnimationBoxQuery by attaching to an actor
and listening for changes in the actor geometry, updating the
underlying stored geometry in the box as appropriate. The caller
can then query the geometry directly from the box as opposed to
having to reach all the way into Clutter to get the geometry of
the animated surface.